bypass/requirements.txt
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.10.0
attrs==25.4.0
boto3==1.42.78
botocore==1.42.78
branca==0.8.2
cachetools==6.2.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
coverage==7.13.5
cryptography==46.0.6
distro==1.9.0
docker==7.1.0
docutils==0.22.3
duckdb==1.4.2
ecdsa==0.19.2
fastapi==0.135.2
fastjsonschema==2.21.2
folium==0.20.0
gitingest==0.3.1
google-auth==2.43.0
google-genai==1.52.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.3.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_core==5.9.1
loguru==0.7.3
loro==1.8.2
mangum==0.21.0
marimo==0.18.1
Markdown==3.10
MarkupSafe==3.0.3
moto==5.1.22
msgspec==0.20.0
narwhals==2.12.0
nbformat==5.10.4
numpy==2.3.5
openai==2.8.1
-e git+https://github.com/band0000ball/OpenHub.git@8be4311d3a31557ff67ed04bf633aac182fb9e31#egg=openhub_bypass&subdirectory=bypass
outcome==1.3.0.post0
packaging==25.0
parso==0.8.5
pathspec==0.12.1
platformdirs==4.5.0
playwright==1.58.0
pluggy==1.6.0
polars==1.35.2
polars-runtime-32==1.35.2
psutil==7.1.3
py-partiql-parser==0.6.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.1
Pygments==2.19.2
pymdown-extensions==10.17.2
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
regex==2025.7.34
requests==2.32.5
responses==0.26.0
respx==0.22.0
rpds-py==0.29.0
rsa==4.9.1
ruff==0.14.6
s3transfer==0.16.0
setuptools==82.0.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlglot==28.0.0
sqlglotrs==0.7.3
starlette==0.47.3
tenacity==9.1.2
tiktoken==0.11.0
tomlkit==0.13.3
tqdm==4.67.1
traitlets==5.14.3
trio==0.33.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uv==0.8.13
uvicorn==0.38.0
websockets==15.0.1
Werkzeug==3.1.7
win32_setctime==1.2.0
xmltodict==1.0.4
xyzservices==2025.11.0
